# Add feature benchmarking (utf8/structural density)