This adds the ability to approximate the cost of a given feature (a set of code locked behind an if: UTF-8, or large numbers of structurals) so that you can see the tradeoffs you are making when making changes, and compare those tradeoffs directly.

Run on a skylake machine, this yields the following costs (ns/block):

| Empty | 1-7 structurals | 1-7 structural miss | utf-8 | utf-8 miss | 8-15 structurals | 8-15 structural miss | 23 structurals | 23 structural miss |
|---|---|---|---|---|---|---|---|---|
| 7.5699 ns/block | +1.4367 ns/block | +0.9489 ns/block | +4.5152 ns/block | +1.8317 ns/block | +1.4982 ns/block | +1.1258 ns/block | +1.9513 ns/block | +1.2484 ns/block |

This means, for example, that a file with 1-7 structurals in every 64-byte block, and no utf-8, should spend 7.6144 + 1.4237 = 9.0381 ns/block in stage 1. A file with 7 structurals per block, with 30% of them UTF-8 (estimating a 20% misprediction, possibly too high), would be on the order of 7.6144 + 1.4237 + 4.6409 + (1.8145*20%) = 14.0419 ns/block. Hovering somewhere around twitter, there.

These are the cost-calculated values of three of our jsonexamples files, vs. the actual measured value, on Skylake:

| File | Expected | Actual |
|---|---|---|
| gsoc-2018.json | 8.97091 ns/block | 9.08471 ns/block |
| twitter.json | 14.293 ns/block | 12.9626 ns/block |
| random.json | 21.9231 ns/block | 17.7019 ns/block |

# Benchmarking Features

This adds a new make target, `run_benchfeatures`, which creates and benchmarks designed to measure the cost of each stage 1 feature in nanoseconds per block.

- Each file is 640KB (because [640K ought to be enough for anyone](https://quoteinvestigator.com/2011/09/08/640k-enough/)) for a total of 10,000 64-byte blocks.
- **Baseline**: we measure the baseline with a 640KB file, with a single value in the first block and the rest spaces.
- **Structurals**: we measure the impact of writing structural indexes to tape with files containing 5000 contiguous blocks with n structurals in each block (where n is 7, 15 or 23). The remaining 5000 blocks are all spaces. (See "Branch miss cost" for the reasoning behind this peculiar choice.)
- **UTF-8**: we measure the impact of UTF-8 with 10000 blocks of 7 structurals each, the first 5000 of which have UTF-8, and the remaining 5000 are ASCII. This is then compared with a pure 10000-block ASCII-only 7-structural-block file to isolate the cost of UTF-8.
- **Branch miss cost**: We measure the *branch miss cost* of each feature by comparing its normal time (5000 contiguous blocks with feature on, 5000 with feature off) to a file with the same number of "on" and "off" blocks, just spread randomly across the file to create roughly 2500 "flips"--places where contiguous regions of the feature begin or end. To ensure consistency across runs and comparisons, all files are based on a single pattern of "on" and "off" in `benchmark/miss-patterns` that was generated from a handpicked random seed, designed to create 2500 discontinuities. The difference in performance between these files, divided by 2500, gives the cost per misprediction of the feature. (This assumes the number of mispredictions is around 2500; on Haswell and Skylake, this appears to be the case--see numbers at the bottom of this post. Regardless, these misprediction costs should be comparable across branches even if they aren't directly comparable across processors.)
- **Sanity check**: `run_benchfeatures` also uses these numbers to predict the running time of gsoc-2018, twitter and random.json (by counting the utf8 and n-structural blocks, and estimating # flips == # mispredicts) and compares the estimate to the actual. They are very, very close for gsoc-2018, a little pessimistic for twitter, and very pessimistic for random.json. My ad-hoc checks indicate it's the branch predictor doing better on these real-world files than on my hand-crafted misprediction nightmares.

At the moment, benchfeatures only reports stage 1 information, because that was my focus; but it's using the standard benchmarker and actually does run and record stage 2 results, so that will be pretty easy to add.

# `parse` changes

The benchmarker still does the same job, but I pulled it out into a separate class and made some enhancements:

* stage 1, stage 2 and allocation numbers are collected separately and merged for a "full run" number at the end, so that you can get stage 1 numbers even when you're measuring a full run.
* parse reports stage 1 and stage 2 separately, as well as the combined full run number (which includes allocation as well).
* `parse -t` now reports the stage 1 and stage 2 throughput after the full-run number (so the 6th column is still the full-run throughput)
* It collects more statistics on the file, including # of blocks with utf-8, # of blocks with 0, 1+, 8+ and 16+ structurals, and "flips" of each of these for misprediction estimation.
* Warmups are removed, since we rely on the best time anyway. Now it is able to run exactly one iteration (minimum before was 2).
* It's capable of benchmarking multiple files in one run. It interleaves their iterations to increase the chances of a "good run," using the `-i` parameter to control the number of times a single file is run continiuously, for cache behavior.
* Prints a progress bar (which it erases when done) for long runs
* No longer reports a perf_open error if it can't collect hardware events, which is the bane of my windows existence and also annoying when I'm not root :)
* The benchmarker code is moved out into benchmarker.h.
* The `-f` parameter (stage 1 only) was changed to `-s stage1`
* The `-s` parameter (force SSE4.2) was changed to `-a HASWELL|WESTMERE|ARM64`.

```
jkeiser@skylake:~/simdjson$ ./parse -n 10000 -i 20 jsonexamples/twitter.json jsonexamples/gsoc-2018.json jsonexamples/generated/utf-8.json jsonexamples/generated/utf-8-miss.json 
make: 'parse' is up to date.
number of iterations 10000 
                                                     
jsonexamples/twitter.json
=========================
     9867 blocks -     631515 bytes - 55263 structurals (  8.8 %)
special blocks with: utf8      2284 ( 23.1 %) - 0 structurals      1287 ( 13.0 %) - 1+ structurals      8581 ( 87.0 %) - 8+ structurals      3272 ( 33.2 %) - 16+ structurals         0 (  0.0 %)
special block flips: utf8      1104 ( 11.2 %) - 0 structurals       940 (  9.5 %) - 1+ structurals       940 (  9.5 %) - 8+ structurals      2593 ( 26.3 %) - 16+ structurals         0 (  0.0 %)

All Stages
|    Speed        :  26.6911 ns per block ( 98.5 %) -   0.4171 ns per byte -   4.7661 ns per structural -    2.398 GB/s
|    Cycles       : 98.569 per block (98.61 %) - 1.540 per byte - 17.601 per structural - 3.693 GHz est. frequency
|    Instructions : 316.18 per block (100.00 %) - 4.94 per byte - 56.46 per structural - 3.21 per cycle
|    Misses       : 2923.00 branch misses (87.93 %) - 0.00 cache misses ( 0.00 %) - 83838.00 cache references
|- Stage 1
|    Speed        :  12.9133 ns per block ( 47.7 %) -   0.2018 ns per byte -   2.3058 ns per structural -    4.956 GB/s
|    Cycles       : 47.687 per block (47.71 %) - 0.745 per byte - 8.515 per structural - 3.693 GHz est. frequency
|    Instructions : 165.25 per block (52.26 %) - 2.58 per byte - 29.51 per structural - 3.47 per cycle
|    Misses       : 1327.00 branch misses (39.92 %) - 0.00 cache misses ( 0.00 %) - 29366.00 cache references
|- Stage 2
|    Speed        :  13.7329 ns per block ( 50.7 %) -   0.2146 ns per byte -   2.4522 ns per structural -    4.660 GB/s
|    Cycles       : 50.711 per block (50.73 %) - 0.792 per byte - 9.055 per structural - 3.693 GHz est. frequency
|    Instructions : 150.93 per block (47.74 %) - 2.36 per byte - 26.95 per structural - 2.98 per cycle
|    Misses       : 1490.00 branch misses (44.82 %) - 1.00 cache misses ( 0.08 %) - 54681.00 cache references

jsonexamples/gsoc-2018.json
===========================
    51997 blocks -    3327831 bytes - 75841 structurals (  2.3 %)
special blocks with: utf8         0 (  0.0 %) - 0 structurals     36833 ( 70.8 %) - 1+ structurals     15165 ( 29.2 %) - 8+ structurals      2815 (  5.4 %) - 16+ structurals         0 (  0.0 %)
special block flips: utf8         0 (  0.0 %) - 0 structurals      7800 ( 15.0 %) - 1+ structurals      7800 ( 15.0 %) - 8+ structurals      3621 (  7.0 %) - 16+ structurals         0 (  0.0 %)

All Stages
|    Speed        :  17.7847 ns per block ( 99.2 %) -   0.2779 ns per byte -  12.1935 ns per structural -    3.599 GB/s
|    Cycles       : 65.658 per block (99.32 %) - 1.026 per byte - 45.016 per structural - 3.692 GHz est. frequency
|    Instructions : 182.84 per block (100.00 %) - 2.86 per byte - 125.36 per structural - 2.78 per cycle
|    Misses       : 19977.00 branch misses (96.95 %) - 14366.00 cache misses (75.79 %) - 361702.00 cache references
|- Stage 1
|    Speed        :   9.0771 ns per block ( 50.6 %) -   0.1418 ns per byte -   6.2234 ns per structural -    7.051 GB/s
|    Cycles       : 33.510 per block (50.69 %) - 0.524 per byte - 22.975 per structural - 3.692 GHz est. frequency
|    Instructions : 114.70 per block (62.73 %) - 1.79 per byte - 78.64 per structural - 3.42 per cycle
|    Misses       : 5556.00 branch misses (26.96 %) - 935.00 cache misses ( 4.93 %) - 123240.00 cache references
|- Stage 2
|    Speed        :   8.6854 ns per block ( 48.4 %) -   0.1357 ns per byte -   5.9549 ns per structural -    7.369 GB/s
|    Cycles       : 32.065 per block (48.50 %) - 0.501 per byte - 21.984 per structural - 3.692 GHz est. frequency
|    Instructions : 68.14 per block (37.27 %) - 1.06 per byte - 46.72 per structural - 2.13 per cycle
|    Misses       : 14430.00 branch misses (70.03 %) - 13045.00 cache misses (68.82 %) - 238447.00 cache references

jsonexamples/generated/utf-8.json
=================================
    10000 blocks -     640000 bytes - 70000 structurals ( 10.9 %)
special blocks with: utf8      5000 ( 50.0 %) - 0 structurals         0 (  0.0 %) - 1+ structurals     10000 (100.0 %) - 8+ structurals         0 (  0.0 %) - 16+ structurals         0 (  0.0 %)
special block flips: utf8         2 (  0.0 %) - 0 structurals         0 (  0.0 %) - 1+ structurals         0 (  0.0 %) - 8+ structurals         0 (  0.0 %) - 16+ structurals         0 (  0.0 %)

All Stages
|    Speed        :  26.0797 ns per block ( 99.6 %) -   0.4075 ns per byte -   3.7257 ns per structural -    2.454 GB/s
|    Cycles       : 96.311 per block (99.69 %) - 1.505 per byte - 13.759 per structural - 3.693 GHz est. frequency
|    Instructions : 361.82 per block (100.00 %) - 5.65 per byte - 51.69 per structural - 3.76 per cycle
|    Misses       : 13.00 branch misses (35.96 %) - 14.00 cache misses ( 1.36 %) - 78480.00 cache references
|- Stage 1
|    Speed        :  13.5119 ns per block ( 51.6 %) -   0.2111 ns per byte -   1.9303 ns per structural -    4.737 GB/s
|    Cycles       : 49.898 per block (51.65 %) - 0.780 per byte - 7.128 per structural - 3.693 GHz est. frequency
|    Instructions : 184.79 per block (51.07 %) - 2.89 per byte - 26.40 per structural - 3.70 per cycle
|    Misses       : 7.00 branch misses (19.36 %) - 0.00 cache misses ( 0.00 %) - 31720.00 cache references
|- Stage 2
|    Speed        :  12.5512 ns per block ( 47.9 %) -   0.1961 ns per byte -   1.7930 ns per structural -    5.099 GB/s
|    Cycles       : 46.348 per block (47.98 %) - 0.724 per byte - 6.621 per structural - 3.693 GHz est. frequency
|    Instructions : 177.03 per block (48.93 %) - 2.77 per byte - 25.29 per structural - 3.82 per cycle
|    Misses       : 4.00 branch misses (11.06 %) - 0.00 cache misses ( 0.00 %) - 46728.00 cache references

jsonexamples/generated/utf-8-miss.json
======================================
    10000 blocks -     640000 bytes - 70000 structurals ( 10.9 %)
special blocks with: utf8      5000 ( 50.0 %) - 0 structurals         0 (  0.0 %) - 1+ structurals     10000 (100.0 %) - 8+ structurals         0 (  0.0 %) - 16+ structurals         0 (  0.0 %)
special block flips: utf8      2500 ( 25.0 %) - 0 structurals         0 (  0.0 %) - 1+ structurals         0 (  0.0 %) - 8+ structurals         0 (  0.0 %) - 16+ structurals         0 (  0.0 %)

All Stages
|    Speed        :  27.9096 ns per block ( 99.3 %) -   0.4361 ns per byte -   3.9871 ns per structural -    2.293 GB/s
|    Cycles       : 103.061 per block (99.34 %) - 1.610 per byte - 14.723 per structural - 3.693 GHz est. frequency
|    Instructions : 361.82 per block (100.00 %) - 5.65 per byte - 51.69 per structural - 3.51 per cycle
|    Misses       : 2240.00 branch misses (92.38 %) - 0.00 cache misses ( 0.00 %) - 78434.00 cache references
|- Stage 1
|    Speed        :  15.3402 ns per block ( 54.6 %) -   0.2397 ns per byte -   2.1915 ns per structural -    4.172 GB/s
|    Cycles       : 56.644 per block (54.60 %) - 0.885 per byte - 8.092 per structural - 3.693 GHz est. frequency
|    Instructions : 184.79 per block (51.07 %) - 2.89 per byte - 26.40 per structural - 3.26 per cycle
|    Misses       : 2235.00 branch misses (92.17 %) - 0.00 cache misses ( 0.00 %) - 31664.00 cache references
|- Stage 2
|    Speed        :  12.5343 ns per block ( 44.6 %) -   0.1958 ns per byte -   1.7906 ns per structural -    5.106 GB/s
|    Cycles       : 46.290 per block (44.62 %) - 0.723 per byte - 6.613 per structural - 3.693 GHz est. frequency
|    Instructions : 177.03 per block (48.93 %) - 2.77 per byte - 25.29 per structural - 3.82 per cycle
|    Misses       : 5.00 branch misses ( 0.21 %) - 0.00 cache misses ( 0.00 %) - 46731.00 cache references
```