@jkeiser 2019-12-17
One interesting learning: cache misses have an outsized effect on branch misses (either rate or cost)--running with -i 1 vs. -i 20, which causes tons of cache misses, increases the branch miss feature costs by 2-3x while most other numbers stay reasonably close to their old values.