@jkeiser 2019-12-17
For a real-world comparison, here is `master` vs. `lookup2`:

| Branch | Empty | 1-7 structurals | 1-7 structural miss | utf-8 | utf-8 miss | 8-15 structurals | 8-15 structural miss | 23 structurals | 23 structural miss |
|---|---|---|---|---|---|---|---|---|---|
| `master` | 7.5699 ns/block | +1.4367 ns/block | +0.9489 ns/block | +4.5152 ns/block | +1.8317 ns/block | +1.4982 ns/block | +1.1258 ns/block | +1.9513 ns/block | +1.2484 ns/block |
| `lookup2` | 7.5442 ns/block | +1.4602 ns/block | +0.8803 ns/block | +4.171 ns/block | +0.9613 ns/block | +1.4763 ns/block | +1.0904 ns/block | +2.2986 ns/block | +1.1661 ns/block |