@lemire 2019-12-17
Merge at will. 

> cache misses have an outsized effect on branch misses

Hmmm...? I am not sure how this can be?