@lemire 2019-12-17
> cache misses have an outsized effect on branch misses

I could see how branch misses might cause cache misses...